fix renamed usecase
