+++ app/src/main/java/org/dhis2/usescases/main/domain/LogoutUser.kt
@@ -3,28 +3,29 @@ package org.dhis2.usescases.main.domain
 import org.dhis2.commons.filters.FilterManager
 import org.dhis2.data.service.SyncStatusController
 import org.dhis2.data.service.workManager.WorkManagerController
-import org.dhis2.mobile.commons.domain.UseCase
-import org.dhis2.mobile.commons.error.DomainError
-import org.dhis2.usescases.main.data.HomeRepository
+import org.dhis2.usescases.main.HomeRepository
 
 typealias AccountCount = Int
 
 class LogoutUser(
+    private val repository: HomeRepository,
     private val workManagerController: WorkManagerController,
     private val syncStatusController: SyncStatusController,
     private val filterManager: FilterManager,
-    private val repository: HomeRepository,
-) : UseCase<Unit, AccountCount> {
-    override suspend fun invoke(input: Unit): Result<AccountCount> =
-        try {
-            workManagerController.cancelAllWork()
-            syncStatusController.restore()
-            filterManager.clearAllFilters()
-            repository.clearPin()
-            repository.logOut()
-            val accountCount = repository.accountsCount()
-            Result.success(accountCount)
-        } catch (domainError: DomainError) {
-            Result.failure(domainError)
-        }
+) {
+    suspend operator fun invoke(): Result<AccountCount> {
+        workManagerController.cancelAllWork()
+        syncStatusController.restore()
+        filterManager.clearAllFilters()
+
+        repository
+            .clearSessionLock()
+            .onFailure { return Result.failure(it) }
+
+        repository
+            .logOut()
+            .onFailure { return Result.failure(it) }
+
+        return Result.success(repository.accountsCount())
+    }
 }
